+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Braindrop ChangeLog
 
+## Unreleased
+
+**Released: 2025-10-21**
+
+- Added `--theme` as a command line switch.
+  ([#190](https://github.com/davep/braindrop/pull/190))
+
 ## v0.12.0
 
 **Released: 2025-10-21**