+++ ChangeLog.md
@@ -6,6 +6,9 @@
 
 - Added `--theme` as a command line switch.
   ([#190](https://github.com/davep/braindrop/pull/190))
+- Deduplicated tag suggestions in the raindrop editor dialog (so now things
+  like `AI` vs `ai` won't be seen as different suggestions).
+  ([#191](https://github.com/davep/braindrop/pull/191))
 
 ## v0.12.0